# crates/mantissa-rest/openapi/mantissa-rest.openapi.json
{
  "components": {
    "schemas": {
...
      "NetworkAttachment": {
        "description": "REST-facing workload attachment row for one overlay network.",
        "properties": {
          "assigned_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "attachment_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance_id": {
            "type": "string"
          },
          "mac": {
            "type": [
              "string",
              "null"
            ]
          },
          "network_id": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "requested_ip": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "traffic_published": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "attachment_id",
          "task_id",
          "node_id",
          "instance_id",
          "network_id",
          "created_at",
          "updated_at",
          "state",
          "traffic_published"
        ],
        "type": "object"
      },
      "NetworkCreateRequest": {
        "additionalProperties": false,
        "description": "REST request body for creating an overlay network.",
        "properties": {
          "bpf_programs": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "driver": {
            "$ref": "#/components/schemas/NetworkDriver"
          },
          "mtu": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "sealed": {
            "type": "boolean"
          },
          "subnet_cidr": {
            "type": [
              "string",
              "null"
            ]
          },
          "vni": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "driver"
        ],
        "type": "object"
      },
      "NetworkCreateResponse": {
        "description": "REST response returned after creating one network.",
        "properties": {
          "network_id": {
            "type": "string"
          }
        },
        "required": [
          "network_id"
        ],
        "type": "object"
      },
      "NetworkDeleteResponse": {
        "description": "REST response returned after deleting networks.",
        "properties": {
          "deleted": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "type": "object"
      },
      "NetworkDriver": {
        "description": "Networking driver supported by the orchestrator.",
        "enum": [
          "vxlan",
          "bridge"
        ],
        "type": "string"
      },
      "NetworkInspect": {
        "description": "REST-facing network inspection response.",
        "properties": {
          "attachment_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "peers": {
            "items": {
              "$ref": "#/components/schemas/NetworkPeerStatus"
            },
            "type": "array"
          },
          "spec": {
            "$ref": "#/components/schemas/NetworkSpec"
          }
        },
        "required": [
          "spec",
          "peers",
          "attachment_count"
        ],
        "type": "object"
      },
...
      "NetworkPeerStatus": {
        "description": "REST-facing network peer convergence row.",
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "peer_id": {
            "type": "string"
          },
          "peer_name": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "peer_id",
          "peer_name",
          "state",
          "updated_at"
        ],
        "type": "object"
      },
      "NetworkSpec": {
        "description": "REST-facing canonical network specification.",
        "properties": {
          "bpf_programs": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "driver": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "mtu": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "sealed": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "subnet_cidr": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "vni": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "driver",
          "subnet_cidr",
          "vni",
          "mtu",
          "created_at",
          "updated_at",
          "status",
          "sealed",
          "bpf_programs"
        ],
        "type": "object"
      },
      "NetworkSummary": {
        "description": "REST-facing network summary row.",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "driver": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "peer_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "ready_peers": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "subnet_cidr": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "vni": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "driver",
          "status",
          "vni",
          "subnet_cidr",
          "peer_count",
          "ready_peers",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "RestErrorBody": {
        "description": "Stable JSON error body returned by REST handlers.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Mantissa REST token",
        "description": "Daemon-generated REST token presented as a Bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "mtls": {
        "description": "Required when the REST listener is bound to a non-loopback interface.",
        "type": "mutualTLS"
      }
    }
  },
  "info": {
    "description": "Local HTTP facade over the Mantissa Cap'n Proto admin session. All /v1 endpoints require an Authorization: Bearer token. Non-loopback listeners must also be served with TLS and client certificate verification enabled.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Mantissa REST API",
    "version": "0.1.0"
  },
  "openapi": "3.2.0",
  "paths": {
...
    "/v1/networks": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Overlay networks visible to the local daemon."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists overlay networks visible to the local daemon.",
        "tags": [
          "networks"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkCreateResponse"
                }
              }
            },
            "description": "Created network identifier."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Creates one overlay network through the local daemon.",
        "tags": [
          "networks"
        ]
      }
    },
    "/v1/networks/{network_id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "description": "Network UUID string.",
            "in": "path",
            "name": "network_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDeleteResponse"
                }
              }
            },
            "description": "Deleted network count."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Deletes one overlay network by UUID string.",
        "tags": [
          "networks"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "description": "Network UUID string.",
            "in": "path",
            "name": "network_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkInspect"
                }
              }
            },
            "description": "Network inspection response."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches one overlay network inspection by UUID string.",
        "tags": [
          "networks"
        ]
      }
    },
    "/v1/networks/{network_id}/attachments": {
      "get": {
        "operationId": "attachments",
        "parameters": [
          {
            "description": "Network UUID string.",
            "in": "path",
            "name": "network_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkAttachment"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Network workload attachment rows."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists workload attachment rows for one overlay network.",
        "tags": [
          "networks"
        ]
      }
    },
    "/v1/networks/{network_id}/peers": {
      "get": {
        "operationId": "peers",
        "parameters": [
          {
            "description": "Network UUID string.",
            "in": "path",
            "name": "network_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkPeerStatus"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Network peer convergence rows."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists per-peer convergence rows for one overlay network.",
        "tags": [
          "networks"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}